requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
python-dotenv
